Fix synchronization issues when changing episode
