+++ LibCarla/source/carla/client/detail/Episode.h
@@ -6,9 +6,13 @@
 
 #pragma once
 
+#include "carla/AtomicSharedPtr.h"
 #include "carla/NonCopyable.h"
+#include "carla/RecurrentSharedFuture.h"
+#include "carla/client/Timestamp.h"
 #include "carla/client/detail/CachedActorList.h"
-#include "carla/client/detail/Client.h"
+#include "carla/client/detail/CallbackList.h"
+#include "carla/client/detail/EpisodeState.h"
 #include "carla/rpc/EpisodeInfo.h"
 
 namespace carla {
@@ -27,7 +31,11 @@ namespace detail {
       private NonCopyable {
   public:
 
-    explicit Episode(uint64_t id) : _id(id) {}
+    explicit Episode(Client &client);
+
+    ~Episode();
+
+    void Listen();
 
     auto GetId() const {
       return GetState()->GetEpisodeId();
@@ -41,13 +49,14 @@ namespace detail {
       _actors.Insert(std::move(actor));
     }
 
-    template <typename RangeT>
-    std::vector<rpc::Actor> GetActors(Client &client, const RangeT &actor_ids) {
-      auto missing_ids = _actors.GetMissingIds(actor_ids);
-      if (!missing_ids.empty()) {
-        _actors.InsertRange(client.GetActorsById(missing_ids));
-      }
-      return _actors.GetActorsById(actor_ids);
+    std::vector<rpc::Actor> GetActors();
+
+    boost::optional<Timestamp> WaitForState(time_duration timeout) {
+      return _timestamp.WaitFor(timeout);
+    }
+
+    void RegisterOnTickEvent(std::function<void(Timestamp)> callback) {
+      _on_tick_callbacks.RegisterCallback(std::move(callback));
     }
 
   private:
